refactor: initial refactor for versioned inference pipline
