+++ src/edvise/runtime/__init__.py
@@ -1,1 +1,1 @@
-"""Runtime helpers shipped inside the Edvise wheel (e.g. versioned inference driver)."""
+"""Runtime helpers for versioned inference launcher orchestration."""
