+++ src/edvise/runtime/versioned_inference/__init__.py
@@ -1,0 +1,1 @@
+"""Version-safe inference launcher (schema-agnostic; PDP is the first consumer)."""
