+++ src/edvise/runtime/versioned_inference/run_metadata.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 import logging
-import os
 from typing import Any
 
 LOGGER = logging.getLogger(__name__)
@@ -12,32 +11,16 @@ VERSIONED_INFERENCE_LAUNCHER_RUN_TYPE = "versioned_inference_launcher"
 _UNRESOLVED_JOB_RUN_ID = "{{job.run_id}}"
 
 
-def get_databricks_run_id() -> str | None:
-    """
-    Best-effort launcher job run id from the Databricks driver environment.
-
-    Prefer :func:`resolve_launcher_run_id` with the job parameter
-    ``launcher_run_id`` (default ``{{job.run_id}}``) instead of this alone.
-    """
-    for key in ("DATABRICKS_RUN_ID",):
-        raw = os.environ.get(key)
-        if raw is not None and str(raw).strip():
-            return str(raw).strip()
-    return None
-
-
 def resolve_launcher_run_id(cli_value: str | None = None) -> str | None:
     """
-    Resolve the parent launcher job run id.
+    Resolve the parent launcher job run id from the Databricks job parameter.
 
-    Prefer the Databricks job parameter ``{{job.run_id}}`` (passed as
-    ``--launcher_run_id``). Fall back to the driver env only when the CLI value
-    is missing or still an unresolved template.
+    Expect ``launcher_run_id`` (default ``{{job.run_id}}``) to be passed on the CLI.
     """
     raw = (cli_value or "").strip()
     if raw and raw != _UNRESOLVED_JOB_RUN_ID:
         return raw
-    return get_databricks_run_id()
+    return None
 
 
 def record_versioned_inference_launcher_event(
@@ -47,7 +30,7 @@ def record_versioned_inference_launcher_event(
     databricks_institution_name: str,
     model_name: str,
     model_run_id: str | None = None,
-    pipeline_version: str | None = None,
+    archived_pipeline_version: str | None = None,
     launcher_run_id: str | None = None,
     child_inference_run_id: str | int | None = None,
     cohort_dataset_name: str | None = None,
@@ -63,9 +46,14 @@ def record_versioned_inference_launcher_event(
     submitted, ``payload`` records the parent → child link
     (``parent_launcher_run_id``, ``child_inference_run_id``, ``db_run_id``).
 
+    ``archived_pipeline_version`` is the git SHA or release tag from training
+    ``config.toml`` — the ref inference runs at, not the launcher deploy version.
+    The same value is written to the ``pipeline_version`` column for dashboard
+    compatibility.
+
     Best-effort: observability failures must not fail the launcher.
     """
-    run_id = launcher_run_id or get_databricks_run_id()
+    run_id = resolve_launcher_run_id(launcher_run_id)
     if not run_id:
         logger.warning(
             "versioned_inference_launcher: skip pipeline_runs write (no launcher run_id)"
@@ -76,8 +64,9 @@ def record_versioned_inference_launcher_event(
     body.setdefault("launcher_job", "edvise_versioned_inference_launcher")
     body.setdefault("model_name", model_name)
     body.setdefault("parent_launcher_run_id", str(run_id))
-    # Child inference silver tables use parent launcher run id as db_run_id.
     body.setdefault("db_run_id", str(run_id))
+    if archived_pipeline_version is not None:
+        body.setdefault("archived_pipeline_version", archived_pipeline_version)
     if child_inference_run_id is not None:
         body["child_inference_run_id"] = str(child_inference_run_id)
 
@@ -100,7 +89,7 @@ def record_versioned_inference_launcher_event(
         cohort_dataset_name=cohort_dataset_name,
         course_dataset_name=course_dataset_name,
         model_run_id=model_run_id,
-        pipeline_version=pipeline_version,
+        pipeline_version=archived_pipeline_version,
         error_message=error_message,
         payload=body,
     )
