+++ src/edvise/runtime/versioned_inference/submit.py
@@ -17,16 +17,16 @@ from datetime import timedelta
 from pathlib import Path
 from typing import Any
 
-from pipelines.pdp.launchers.bundle_from_dab import (
+from edvise.runtime.versioned_inference.bundle.from_dab import (
     DEFAULT_INFERENCE_JOB_KEY,
     DEFAULT_INFERENCE_YML,
     inference_yml_path,
     load_inference_job_definition,
 )
-from pipelines.pdp.launchers.inference_parameters import (
+from edvise.runtime.versioned_inference.parameters import (
     resolve_versioned_job_parameters,
 )
-from pipelines.pdp.launchers.pipeline_version_ref import build_git_source, git_ref_kind
+from edvise.runtime.versioned_inference.pipeline_version_ref import build_git_source, git_ref_kind
 
 LOGGER = logging.getLogger(__name__)
