+++ src/edvise/runtime/versioned_inference/tasks/trigger.py
@@ -1,71 +1,32 @@
-#!/usr/bin/env python3
-"""
-Databricks task 3: submit the full PDP inference pipeline from the archived bundle.
-
-Reads materialized YAML at ``pipeline_version`` (Git SHA in dev, release tag in staging),
-submits a multi-task Jobs API run pinned to that ref, and polls until completion.
-Fails closed — no fallback to current-deploy inference.
-"""
+"""Task 3: submit versioned inference from archived bundle."""
 
 from __future__ import annotations
 
 import argparse
-import inspect
 import json
 import logging
 import sys
-from pathlib import Path
 
-
-def _setup_import_path() -> None:
-    try:
-        launcher = Path(__file__).resolve().parent
-    except NameError:
-        launcher = None
-        for frame_info in inspect.stack():
-            fn = frame_info.filename
-            if not fn or fn.startswith("<"):
-                continue
-            path = Path(fn).resolve()
-            if path.parent.name == "launchers" and path.parent.parent.name == "pdp":
-                launcher = path.parent
-                break
-        if launcher is None:
-            candidate = Path.cwd() / "pipelines" / "pdp" / "launchers"
-            if candidate.is_dir():
-                launcher = candidate.resolve()
-        if launcher is None:
-            msg = "Cannot locate pipelines/pdp/launchers (Databricks import bootstrap)"
-            raise RuntimeError(msg)
-    launcher_str = str(launcher.resolve())
-    if launcher_str not in sys.path:
-        sys.path.insert(0, launcher_str)
-    from _paths import ensure_repo_root_on_sys_path  # noqa: WPS433
-
-    ensure_repo_root_on_sys_path()
-
-
-_setup_import_path()
-
-from pipelines.pdp.launchers.inference_job_submit import (  # noqa: E402
+from edvise.runtime.versioned_inference.submit import (
     DEFAULT_GIT_URL,
     submit_versioned_inference_from_bundle,
 )
-from pipelines.pdp.launchers.launcher_cli import (  # noqa: E402
+from edvise.runtime.versioned_inference.cli import (
     add_inference_trigger_args,
     build_launcher_trigger_inputs,
     optional_model_run_id,
 )
-from pipelines.pdp.launchers.launcher_run_metadata import (  # noqa: E402
-    resolve_launcher_run_id,
-    record_versioned_inference_launcher_event,
-)
-from pipelines.pdp.launchers.model_metadata import (  # noqa: E402
+from edvise.runtime.versioned_inference.model_resolution import (
     get_spark_session,
     resolve_model_run_and_pipeline_version,
     resolve_release_dir,
 )
-from pipelines.pdp.launchers.pipeline_version_ref import git_ref_kind  # noqa: E402
+from edvise.runtime.versioned_inference.pipeline_version_ref import git_ref_kind
+from edvise.runtime.versioned_inference.dab_layout import resolve_dab_bundle_layout
+from edvise.runtime.versioned_inference.run_metadata import (
+    record_versioned_inference_launcher_event,
+    resolve_launcher_run_id,
+)
 
 LOGGER = logging.getLogger("trigger_versioned_inference")
 
@@ -103,7 +64,7 @@ def _fail(
     inst: str,
     model: str,
     model_run_id: str | None,
-    pipeline_version: str | None,
+    archived_pipeline_version: str | None,
     launcher_run_id: str | None,
     message: str,
 ) -> int:
@@ -114,7 +75,7 @@ def _fail(
         databricks_institution_name=inst,
         model_name=model,
         model_run_id=model_run_id,
-        pipeline_version=pipeline_version,
+        archived_pipeline_version=archived_pipeline_version,
         launcher_run_id=launcher_run_id,
         error_message=message,
         payload={"task": "trigger_versioned_inference"},
@@ -146,11 +107,11 @@ def main(argv: list[str] | None = None) -> int:
             inst=inst,
             model=model,
             model_run_id=None,
-            pipeline_version=None,
+            archived_pipeline_version=None,
             launcher_run_id=None,
             message=(
                 "launcher_run_id is required (job parameter launcher_run_id with "
-                "default {{job.run_id}}). Synthetic versioned_* db_run_id is not used."
+                "default {{job.run_id}})."
             ),
         )
     try:
@@ -161,7 +122,7 @@ def main(argv: list[str] | None = None) -> int:
             inst=inst,
             model=model,
             model_run_id=None,
-            pipeline_version=None,
+            archived_pipeline_version=None,
             launcher_run_id=launcher_run_id,
             message=f"Invalid inference parameter overrides: {exc}",
         )
@@ -173,7 +134,7 @@ def main(argv: list[str] | None = None) -> int:
             inst=inst,
             model=model,
             model_run_id=None,
-            pipeline_version=None,
+            archived_pipeline_version=None,
             launcher_run_id=launcher_run_id,
             message="SparkSession is required (run on Databricks).",
         )
@@ -192,26 +153,27 @@ def main(argv: list[str] | None = None) -> int:
             inst=inst,
             model=model,
             model_run_id=None,
-            pipeline_version=None,
+            archived_pipeline_version=None,
             launcher_run_id=launcher_run_id,
             message="Could not resolve model_run_id / pipeline_version",
         )
-    model_run_id, pipeline_version = resolved
+    model_run_id, archived_pipeline_version = resolved
+    layout = resolve_dab_bundle_layout(inputs.schema_type)
     LOGGER.info(
-        "Triggering inference for model_run_id=%s pipeline_version=%s (git %s)",
+        "Triggering inference for model_run_id=%s archived_pipeline_version=%s (git %s)",
         model_run_id,
-        pipeline_version,
-        git_ref_kind(pipeline_version),
+        archived_pipeline_version,
+        git_ref_kind(archived_pipeline_version),
     )
 
-    release_dir = resolve_release_dir(inputs.release_base_path, pipeline_version)
+    release_dir = resolve_release_dir(inputs.release_base_path, archived_pipeline_version)
     if not release_dir.is_dir():
         return _fail(
             catalog=db_ws,
             inst=inst,
             model=model,
             model_run_id=model_run_id,
-            pipeline_version=pipeline_version,
+            archived_pipeline_version=archived_pipeline_version,
             launcher_run_id=launcher_run_id,
             message=(
                 f"Release bundle not found: {release_dir} "
@@ -222,11 +184,13 @@ def main(argv: list[str] | None = None) -> int:
     try:
         run_id = submit_versioned_inference_from_bundle(
             release_dir,
-            pipeline_version=pipeline_version,
+            archived_pipeline_version=archived_pipeline_version,
             parameter_overrides=inputs.param_overrides,
             extra_parameter_overrides=inputs.extra_param_overrides,
             stable_trigger=inputs.stable_trigger,
             git_url=inputs.git_url,
+            inference_job_key=layout.inference_job_key,
+            inference_yml_relative=layout.inference_yml_snapshot_rel,
             dry_run=args.dry_run,
             wait_for_completion=not args.no_wait,
             poll_interval_seconds=args.poll_interval_seconds,
@@ -238,7 +202,7 @@ def main(argv: list[str] | None = None) -> int:
             inst=inst,
             model=model,
             model_run_id=model_run_id,
-            pipeline_version=pipeline_version,
+            archived_pipeline_version=archived_pipeline_version,
             launcher_run_id=launcher_run_id,
             message=f"Failed to submit or complete versioned inference: {exc}",
         )
@@ -253,7 +217,7 @@ def main(argv: list[str] | None = None) -> int:
             databricks_institution_name=inst,
             model_name=model,
             model_run_id=model_run_id,
-            pipeline_version=pipeline_version,
+            archived_pipeline_version=archived_pipeline_version,
             launcher_run_id=launcher_run_id,
             child_inference_run_id=run_id,
             cohort_dataset_name=inputs.param_overrides.get("cohort_file_name"),
@@ -286,9 +250,3 @@ def main(argv: list[str] | None = None) -> int:
                 db_run_id,
             )
     return 0
-
-
-if __name__ == "__main__":
-    _exit_code = main()
-    if _exit_code:
-        sys.exit(_exit_code)
