+++ src/edvise/runtime/versioned_inference/trigger_contract.py
@@ -21,6 +21,7 @@ VERSIONED_INFERENCE_LAUNCHER_JOB_PARAMETERS: tuple[str, ...] = (
     "model_name",
     "model_run_id",
     "DB_workspace",
+    "schema_type",
     "release_base_path",
     "github_repo",
     "git_url",
@@ -41,6 +42,7 @@ def build_versioned_inference_job_parameters(
     databricks_institution_name: str,
     model_name: str,
     db_workspace: str,
+    schema_type: str = "pdp",
     cohort_file_name: str = "",
     course_file_name: str = "",
     gcp_bucket_name: str = "",
@@ -68,6 +70,7 @@ def build_versioned_inference_job_parameters(
         "model_name": model_name.strip(),
         "model_run_id": model_run_id.strip(),
         "DB_workspace": db_workspace.strip(),
+        "schema_type": schema_type.strip() or "pdp",
         "release_base_path": release_base_path.strip(),
         "github_repo": github_repo.strip(),
         "git_url": git_url.strip(),
