+++ src/edvise/scripts/inference_launcher_validate.py
@@ -1,0 +1,16 @@
+#!/usr/bin/env python3
+"""Launcher task: submit versioned inference from archived bundle YAML."""
+
+from __future__ import annotations
+
+import sys
+from pathlib import Path
+
+_src = Path(__file__).resolve().parents[2]
+if _src.is_dir() and str(_src) not in sys.path:
+    sys.path.insert(0, str(_src))
+
+from edvise.runtime.versioned_inference.tasks.validate import main  # noqa: E402
+
+if __name__ == "__main__":
+    raise SystemExit(main())
