+++ tests/pdp/test_launcher_run_metadata.py
@@ -2,15 +2,10 @@
 
 from __future__ import annotations
 
-import sys
-from pathlib import Path
+from unittest.mock import patch
 
-_REPO_ROOT = Path(__file__).resolve().parents[2]
-if str(_REPO_ROOT) not in sys.path:
-    sys.path.insert(0, str(_REPO_ROOT))
-
-from pipelines.pdp.launchers.launcher_run_metadata import (
-    get_databricks_run_id,
+from edvise.runtime.versioned_inference.run_metadata import (
+    record_versioned_inference_launcher_event,
     resolve_launcher_run_id,
 )
 
@@ -19,16 +14,31 @@ def test_resolve_launcher_run_id_prefers_job_parameter() -> None:
     assert resolve_launcher_run_id("439619245566927") == "439619245566927"
 
 
-def test_resolve_launcher_run_id_ignores_unresolved_template(
-    monkeypatch,
-) -> None:
-    monkeypatch.setenv("DATABRICKS_RUN_ID", "111222333")
-    assert resolve_launcher_run_id("{{job.run_id}}") == "111222333"
-    assert resolve_launcher_run_id("") == "111222333"
+def test_resolve_launcher_run_id_ignores_unresolved_template() -> None:
+    assert resolve_launcher_run_id("{{job.run_id}}") is None
 
 
-def test_resolve_launcher_run_id_none_when_missing(monkeypatch) -> None:
-    monkeypatch.delenv("DATABRICKS_RUN_ID", raising=False)
+def test_resolve_launcher_run_id_none_when_missing() -> None:
     assert resolve_launcher_run_id("") is None
     assert resolve_launcher_run_id("{{job.run_id}}") is None
-    assert get_databricks_run_id() is None
+
+
+def test_record_launcher_event_uses_archived_pipeline_version() -> None:
+    with patch(
+        "edvise.shared.dashboard_metadata.pipeline_runs.append_pipeline_run_event",
+        return_value=True,
+    ) as append:
+        ok = record_versioned_inference_launcher_event(
+            catalog="dev_sst_02",
+            event="completed",
+            databricks_institution_name="midway",
+            model_name="retention",
+            model_run_id="train-123",
+            archived_pipeline_version="abc123def456",
+            launcher_run_id="439619245566927",
+        )
+    assert ok is True
+    append.assert_called_once()
+    kwargs = append.call_args.kwargs
+    assert kwargs["pipeline_version"] == "abc123def456"
+    assert kwargs["payload"]["archived_pipeline_version"] == "abc123def456"
