+++ tests/pdp/test_pipeline_version_ref.py
@@ -9,7 +9,7 @@ _REPO_ROOT = Path(__file__).resolve().parents[2]
 if str(_REPO_ROOT) not in sys.path:
     sys.path.insert(0, str(_REPO_ROOT))
 
-from pipelines.pdp.launchers import pipeline_version_ref as pvr
+from edvise.runtime.versioned_inference import pipeline_version_ref as pvr
 
 
 def test_is_git_commit_sha_full_and_short() -> None:
